Feature: Add tile manipulation functions to Canvas Feature: Allow active palette tab to differ from active image tab Feature: Stub out the toolbox Feature: Add color swatch handling Change: Rename some editor tabs to image tabs Change: Add some warning cleanups (Mostly add override)

+++ tile_data.h
@@ -4,8 +4,8 @@
 #include <QImage>
 
 struct tile_data{
-	short palette;
-	short subpalette;
+	int palette;
+	int subpalette;
 	QImage image;
 };
